refactor: split game_hud.rs into submodules
split 3904-line monolithic file into 7 files under ui/game_hud/:
- mod.rs: top-level hud, resource bar, shared types (861 lines)
- inspector.rs: bottom panel system, tower upgrade, shared types (561 lines)
- building_inspector.rs: building info, mine assignment (816 lines)
- npc_inspector.rs: dc group inspector, npc detail view (885 lines)
- combat_log.rs: log window, filter state (382 lines)
- squad_overlay.rs: squad/target/faction overlays (468 lines)
- build_ghost.rs: placeholder for future build preview

zero behavior change, all callers compile unchanged
