+++ rust/src/ui/game_hud/build_ghost.rs
@@ -1,0 +1,3 @@
+//! Building placement ghost / preview overlay.
+//! The core build ghost rendering lives in `ui/build_menu.rs`.
+//! This module is reserved for HUD-level ghost overlays if needed in future.
